Trait webb_relayer_store::EncryptedOutputCacheStore

pub trait EncryptedOutputCacheStore: HistoryStore {
    type Output: IntoIterator<Item = Vec<u8>>;

    fn get_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K
    ) -> Result<Self::Output>;
    fn get_encrypted_output_with_range<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        range: Range<u32>
    ) -> Result<Self::Output>;
    fn insert_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        encrypted_output: &[(u32, Vec<u8>)]
    ) -> Result<()>;
    fn get_last_deposit_block_number_for_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K
    ) -> Result<u64>;
    fn insert_last_deposit_block_number_for_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        block_number: u64
    ) -> Result<u64>;
}

An Encrypted Output Cache Store is a simple trait that would help in getting the encrypted output and insert them with a simple API.

Required Associated Types

type Output: IntoIterator<Item = Vec<u8>>

The Output type which is the encrypted output.

Required Methods

fn get_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K
) -> Result<Self::Output>

Get all the encrypted output for the given key.

fn get_encrypted_output_with_range<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
range: Range<u32>
) -> Result<Self::Output>

Get the encrypted output for the given key and with the given range of indices.

fn insert_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
encrypted_output: &[(u32, Vec<u8>)]
) -> Result<()>

Insert the encrypted output for the given key.

fn get_last_deposit_block_number_for_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K
) -> Result<u64>

The last deposit info is sent to the client on encrypted_output request So they can verify when the last transaction was sent to maintain their own state of mixers.

fn insert_last_deposit_block_number_for_encrypted_output<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
block_number: u64
) -> Result<u64>

Set the last deposit block number for the given key.