Trait webb_relayer_store::HistoryStore

pub trait HistoryStore: Clone + Send + Sync {
    // Required methods
    fn set_last_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        block_number: u64
    ) -> Result<u64>;
    fn get_last_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        default_block_number: u64
    ) -> Result<u64>;
    fn set_target_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        block_number: u64
    ) -> Result<u64>;
    fn get_target_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K,
        default_block_number: u64
    ) -> Result<u64>;

    // Provided methods
    fn get_last_block_number_or_default<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K
    ) -> Result<u64> { ... }
    fn get_target_block_number_or_default<K: Into<HistoryStoreKey> + Debug>(
        &self,
        key: K
    ) -> Result<u64> { ... }
}

HistoryStore is a simple trait for storing and retrieving history of block numbers.

Required Methods

fn set_last_block_number<K: Into<HistoryStoreKey> + Debug>( &self, key: K, block_number: u64 ) -> Result<u64>

Sets the new block number for that contract in the cache and returns the old one.

fn get_last_block_number<K: Into<HistoryStoreKey> + Debug>( &self, key: K, default_block_number: u64 ) -> Result<u64>

Get the last block number for that contract. if not found, returns the default_block_number.

fn set_target_block_number<K: Into<HistoryStoreKey> + Debug>( &self, key: K, block_number: u64 ) -> Result<u64>

Sets the Target Block number (Usually the latest block number of the target chain) This used to be able to check if we are fully synced with the target chain or not.

fn get_target_block_number<K: Into<HistoryStoreKey> + Debug>( &self, key: K, default_block_number: u64 ) -> Result<u64>

Get the target block number. if not found, returns the default_block_number.

Provided Methods

fn get_last_block_number_or_default<K: Into<HistoryStoreKey> + Debug>( &self, key: K ) -> Result<u64>

an easy way to call the get_last_block_number where the default block number is 1.

fn get_target_block_number_or_default<K: Into<HistoryStoreKey> + Debug>( &self, key: K ) -> Result<u64>

an easy way to call the get_target_block_number.