Trait webb_relayer_store::QueueKey

pub trait QueueKey {
    // Required methods
    fn queue_name(&self) -> String;
    fn item_key(&self) -> Option<[u8; 64]>;
}

A trait for retrieving queue keys

Required Methods

fn queue_name(&self) -> String

The Queue name, used as a prefix for the keys.

fn item_key(&self) -> Option<[u8; 64]>

an optional different key for the same value stored in the queue.

This useful for the case when you want to have a direct access to a specific key in the queue. For example, if you want to remove an item from the queue, you can use this key to directly remove it from the queue.