Crate webb_relayer_store

Relayer Store Module 🕸️

A module for managing the storage of the relayer.

Overview

The relayer store module stores the history of events. Manages the setting and retrieving operations of events.

Re-exports

pub use self::sled::SledStore;

pub use mem::InMemoryStore;

Modules

mem

A module for managing in-memory storage of the relayer.

sled

A module for setting up and managing a Sled-based database.

Structs

BridgeKey

A Bridge Key is a unique key used for Sending and Receiving Commands to the Signature Bridge It is a combination of the Chain ID and the target system of the Bridge system.

Enums

BridgeCommand

A Command sent to the Bridge to execute different actions.

HistoryStoreKey

HistoryStoreKey contains the keys used to store the history of events.

Traits

EncryptedOutputCacheStore

An Encrypted Output Cache Store is a simple trait that would help in getting the encrypted output and insert them with a simple API.

EventHashStore

A Simple Event Store, that does not store the events, instead it store the hash of the event as the key and the value is just empty bytes.

HistoryStore

HistoryStore is a simple trait for storing and retrieving history of block numbers.

IntoTypedChainId

A way to convert an arbitrary type to a TypedChainId.

LeafCacheStore

A Leaf Cache Store is a simple trait that would help in getting the leaves and insert them with a simple API.

QueueKey

A trait for retrieving queue keys

QueueStore

A Queue Store is a simple trait that help storing items in a queue. The queue is a FIFO queue, that can be used to store anything that can be serialized.

TokenPriceCacheStore

A trait for Cached Token Price.