Trait webb_relayer_store::QueueStore

pub trait QueueStore<Item>where
    Item: Serialize + DeserializeOwned + Clone,{
    type Key: QueueKey;

    // Required methods
    fn enqueue_item(&self, key: Self::Key, item: Item) -> Result<()>;
    fn dequeue_item(&self, key: Self::Key) -> Result<Option<Item>>;
    fn peek_item(&self, key: Self::Key) -> Result<Option<Item>>;
    fn has_item(&self, key: Self::Key) -> Result<bool>;
    fn remove_item(&self, key: Self::Key) -> Result<Option<Item>>;
}

A Queue Store is a simple trait that help storing items in a queue. The queue is a FIFO queue, that can be used to store anything that can be serialized.

There is a simple API to get the items from the queue, from a background task for example.

Required Associated Types

type Key: QueueKey

The type of the queue key.

Required Methods

fn enqueue_item(&self, key: Self::Key, item: Item) -> Result<()>

Insert an item into the queue.

fn dequeue_item(&self, key: Self::Key) -> Result<Option<Item>>

Get an item from the queue, and removes it.

fn peek_item(&self, key: Self::Key) -> Result<Option<Item>>

Get an item from the queue, without removing it.

fn has_item(&self, key: Self::Key) -> Result<bool>

Check if the item is in the queue.

fn remove_item(&self, key: Self::Key) -> Result<Option<Item>>

Remove an item from the queue.